Teardown everything and restart when a new HAR is uploaded
When a new HAR is PUT to /har, we need to tear down the old environment and
create a new one for the new HAR.

We have to generate new server certs, as it might be the case that a different
set of origins resolve to a given IP. Unfortunately, Chrome will not connect to
an origin replying with a new cert if it has an existing one. So, we have to
bounce Chrome every time we tear down and initialize a new environment.

Fix #8