Refactor to support multiple origins per IP
The HAR server was only capable of serving a single origin, but two origins
(example.com and www.example.com, e.g.) might resolve to the same IP address.
So, refactor to add support for this scenario. The most substantial result is
that the servers that actually replay the responses don't receive a HAR anymore,
so rename them to `replay-server`.

Fix #16