Rewrite datetime headers relative to request time
Previously, we replayed headers unmodified from the value in the HAR. However,
this isn't what one wants in the case of `Expires`, for example, as that value
is valid relative to when the HAR was *captured*, not when it is replayed.

So, generate the value of all valid HTTP date/time headers by calculating the
delta between the request/response entry's `startedDateTime` and the header
value and then generating a new RFC 1123 date by applying that delta to the when
the replay request was received. Note that all replayed date/time header values
are RFC 1123 dates, even if the original value was in some other format.

If the header was calculated relative to some other event, then the replayed
value may be incorrect. For example, if the original `Expires` header for a
script was calculated relative to when its including document was served, then
it's unlikely that the replayed value will be correct, as the latency between
the two requests probably has changed.

Fix #12