Use httpVersion specified in HAR
The replay server accepts an `--http2` command-line argument that determines
whether HTTPS origins are served over HTTP/1.1 or /2. Now that there may be
multiple origins per server, this is not correct. Instead, use the `httpVersion`
specified in the first response for an instance (<protocol, ip, port> tuple,
i.e.).

Fix #20